Check for download status in get_cpm.cmake
If the download fails, file(DOWNLOAD) will create an empty CPM.cmake
file. This means that successive runs of CMake and get_cpm.cmake
could mistake this file for a proper CPM.cmake script.
To avoid this issue the empty file is deleted before CMake is terminated
with a FATAL_ERROR.